# Conflicts:
#	app/src/main/res/layout/activity_moneysend.xml
#	app/src/main/res/layout/activity_work_out.xml

+++ app/src/main/java/com/example/baehyeonbin/neverending_nightmare/beans/DetailRoom.kt
@@ -1,3 +1,3 @@
 package com.example.baehyeonbin.neverending_nightmare.beans
 
-data class DetailRoom(val result : Room, val balance : Int)
+data class DetailRoom(val result : Room, val balance : Double)
